+++ test/JsonMessageSourceTest.php
@@ -4,12 +4,91 @@ namespace yii\i18n;
 
 use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
-use yii\console\{Application};
 
 /**
  * Tests the features of the `yii\i18n\JsonMessageSource` class.
  */
 class JsonMessageSourceTest extends TestCase {
+
+  /**
+   * @test JsonMessageSource::flatten
+   */
+  public function testFlatten(): void {
+    $flatten = function($array) {
+      return $this->flatten($array);
+    };
+
+    it('should merge the keys of a multidimensional array', function() use ($flatten) {
+      $model = new JsonMessageSource;
+      expect($flatten->call($model, []))->to->equal([]);
+      expect($flatten->call($model, ['foo' => 'bar', 'baz' => 'qux']))->to->equal(['foo' => 'bar', 'baz' => 'qux']);
+      expect($flatten->call($model, ['foo' => ['bar' => 'baz']]))->to->equal(['foo.bar' => 'baz']);
+
+      $source = [
+        'foo' => 'bar',
+        'bar' => ['baz' => 'qux'],
+        'baz' => ['qux' => [
+          'foo' => 'bar',
+          'bar' => 'baz'
+        ]]
+      ];
+
+      expect($flatten->call($model, $source))->to->equal([
+        'foo' => 'bar',
+        'bar.baz' => 'qux',
+        'baz.qux.foo' => 'bar',
+        'baz.qux.bar' => 'baz'
+      ]);
+    });
+
+    it('should allow different nesting separators', function() use ($flatten) {
+      $source = [
+        'foo' => 'bar',
+        'bar' => ['baz' => 'qux'],
+        'baz' => ['qux' => [
+          'foo' => 'bar',
+          'bar' => 'baz'
+        ]]
+      ];
+
+      $model = new JsonMessageSource(['nestingSeparator' => '/']);
+      expect($flatten->call($model, $source))->to->equal([
+        'foo' => 'bar',
+        'bar/baz' => 'qux',
+        'baz/qux/foo' => 'bar',
+        'baz/qux/bar' => 'baz'
+      ]);
+
+      $model = new JsonMessageSource(['nestingSeparator' => '->']);
+      expect($flatten->call($model, $source))->to->equal([
+        'foo' => 'bar',
+        'bar->baz' => 'qux',
+        'baz->qux->foo' => 'bar',
+        'baz->qux->bar' => 'baz'
+      ]);
+    });
+  }
+
+  /**
+   * @test JsonMessageSource::getMessageFilePath
+   */
+  public function testGetMessageFilePath(): void {
+    $getMessageFilePath = function($category, $language) {
+      return $this->getMessageFilePath($category, $language);
+    };
+
+    it('should return the proper path to the message file', function() use ($getMessageFilePath) {
+      $model = new JsonMessageSource(['basePath' => '@root/test/fixtures']);
+      $messageFile = str_replace('/', DIRECTORY_SEPARATOR, __DIR__.'/fixtures/fr/messages.json');
+      expect($getMessageFilePath->call($model, 'messages', 'fr'))->to->equal($messageFile);
+    });
+
+    it('should should support different file extensions', function() use ($getMessageFilePath) {
+      $model = new JsonMessageSource(['basePath' => '@root/test/fixtures', 'fileExtension' => 'json5']);
+      $messageFile = str_replace('/', DIRECTORY_SEPARATOR, __DIR__.'/fixtures/fr/messages');
+      expect($getMessageFilePath->call($model, 'messages', 'fr'))->to->equal("$messageFile.json5");
+    });
+  }
 
   /**
    * @test JsonMessageSource::loadMessagesFromFile
@@ -36,12 +115,17 @@ class JsonMessageSourceTest extends TestCase {
   }
 
   /**
-   * Performs a common set of tasks just before each test method is called.
+   * @test JsonMessageSource::parseMessages
    */
-  protected function setUp(): void {
-    new Application([
-      'id' => 'yii2-i18n-messages',
-      'basePath' => '@root/lib'
-    ]);
+  public function testParseMessages(): void {
+    $parseMessages = function($messageData) {
+      return $this->parseMessages($messageData);
+    };
+
+    // TODO
+    it('should parse a JSON file as a hierarchical array', function() use ($parseMessages) {
+      $model = new JsonMessageSource(['basePath' => '@root/test/fixtures', 'enableNesting' => true]);
+      $messages = $parseMessages->call($model, file_get_contents(\Yii::getAlias("{$model->basePath}/fr/messages.json")));
+    });
   }
 }